Move auth provider related code to own file + use generics when testing success responses
